+++ src/lib/a11y.js
@@ -32,6 +32,5 @@
  * @param {string} name
  * @returns {string}
  */
-const getFieldErrorId = (formId, name) => `${formId}-error-${name}`;
+export const getFieldErrorId = (formId, name) => `${formId}-error-${name}`;
 
-export default getFieldErrorId;