+++ src/lib/core/Context.test.js
@@ -2,10 +2,10 @@ import React from 'react';
 import { fireEvent, render } from '@testing-library/react';
 
 import { ajvSchema } from '../providers/ajv';
-import FormContext, { useFormContext, useResolvedForm, withFormContext } from './Context';
-import Field from './Field';
-import Form from './Form';
-import useForm from './useForm';
+import { FormContext, useFormContext, useResolvedForm, withFormContext } from './Context';
+import { Field } from './Field';
+import { Form } from './Form';
+import { useForm } from './useForm';
 
 const schema = ajvSchema({ type: 'object' });