+++ src/lib/core/Form.test.js
@@ -6,10 +6,10 @@ import {
 } from '@testing-library/react';
 
 import { ajvSchema } from '../providers/ajv';
-import Form from './Form';
-import Field from './Field';
-import FieldError from './FieldError';
-import useForm from './useForm';
+import { Form } from './Form';
+import { Field } from './Field';
+import { FieldError } from './FieldError';
+import { useForm } from './useForm';
 import { useFormContext } from './Context';
 
 const testSchema = {